requirements.txt
decord
einops
imageio
numpy==1.23.5
omegaconf
opencv_python
pandas
Pillow
pytorch_lightning
PyYAML
setuptools
tqdm
transformers
av
gradio
timm
open_clip_torch==2.22.0
kornia
accelerate
diffusers
huggingface_hub
ipykernel
librosa==0.9.2
matplotlib
scikit-learn
scipy
#==1.11.1
soundfile
tokenizers
imageio[ffmpeg]
ninja
wandb
peft
laion-clap==1.1.6
ffmpeg_quality_metrics
pytorchvideo
av
tensorboard
tensorboardX
ffmpeg